Add files via upload
This project is still related to circle, but it gets the area of the circle. The program will ask the user for radius, and then execute the output with the given formula. I forgot to say, but you can also use my projects as a reference if you want to. Thanks.