Add files via upload
This is probably my first project in Java. This project is about getting the circumference of the circle. The program will ask the user for the radius, and then the program will calculate using the given formula, and execute the output. Feel free to use my code as a reference. Thanks.